refactor: Header and Footer components
