+++ frontend/src/app/layout.tsx
@@ -3,15 +3,15 @@
 import type { Metadata } from "next";
 // import { Inter } from "next/font/google";
 import "./globals.css";
-import Footer from "./common/Footer";
+import Footer from "../components/Footer";
 import { Toaster } from "sonner";
 import { Loader } from "@/components/Loader";
 
 // const inter = Inter({ subsets: ["latin"] });
 
 export const metadata: Metadata = {
-  title: "Create Next App",
-  description: "Generated by create next app",
+  title: "AgroSmart",
+  description: "Agricultura Inteligente",
 };
 
 export default function RootLayout({
