+++ frontend/src/app/page.tsx
@@ -1,6 +1,6 @@
 import Link from "next/link";
-import Header from "./common/Header";
-import Footer from "./common/Footer";
+import Header from "../components/Header";
+import Footer from "../components/Footer";
 import withAuth from "./auth/withAuth";
 
 export default function  Home() {
